Move permissions tests to their own file
